feat: UE_4.26 Use new transition API

+++ Source/WebCameraFeed/Private/PixelShaderDeclaration.h
@@ -129,7 +129,7 @@ public:
 		}
 
 
-		SetShaderValue(RHICmdList, ShaderRHI, mirrorParameter, mirror);
+		SetShaderValue(RHICmdList, ShaderRHI, mirrorParameter, (int)mirror);
 		SetTextureParameter(RHICmdList, ShaderRHI, TextureParameter, TextureParameterSampler, samplerState, UITextureRHI->TextureRHI);
 	}
